Remove HAVE_STRING_H
The C89 standard and later defines the <string.h> header as part of the
standard headers [1] and on current systems it is always present.

Code included also <strings.h> header as an alterinative in some files.
This kind of check was relevant on some older systems where the
<strings.h> file included definitions for the C89 compliant <string.h>.
Today such alternative check is not required anymore. The <strings.h>
file is part of the POSIX definition these days. Current code doesn't
require the <strings.h> files in cases of these patched files.

Also Autoconf suggests doing this and relying on C89 or above [2] and [3].

[1]: https://port70.net/~nsz/c/c89/c89-draft.html#4.1.2
[2]: http://git.savannah.gnu.org/cgit/autoconf.git/tree/lib/autoconf/headers.m4
[3]: https://www.gnu.org/software/autoconf/manual/autoconf-2.69/autoconf.html