Remove HAVE_PROTOTYPES and HAVE_STDARG_PROTOTYPES
The C89 standard and later defines the `<stdarg.h>` header as part of
the standard headers [1] and on current systems it is always present.

Checking for presence and functionality of the `<stdarg.h>` header and
variadic functions is not relevant anymore on current systems since this
is always available.

Also Autoconf suggests doing this and relying on C89 or above [2] and [3].

As an alternative, outdated versions of POSIX defined the legacy header
`<varargs.h>` [4].

This patch modernize the `<stdarg.h>` usage a bit.

Unused internal macro `va_init_list()` has been replaced with
`va_start()`.

Checking if functions support prototypes is also not relevant anymore
so the `HAVE_PROTOTYPES` symbol has been removed likewise.

Refs:
[1] https://port70.net/~nsz/c/c89/c89-draft.html#4.1.2
[2] http://git.savannah.gnu.org/cgit/autoconf.git/tree/lib/autoconf/headers.m4
[3] https://www.gnu.org/software/autoconf/manual/autoconf-2.69/autoconf.html
[4] https://en.wikipedia.org/wiki/Stdarg.h#varargs.h