Remove HAVE_STRINGS_H
This patch removes not needed `HAVE_STRINGS_H` symbol as was defined by
Autoconf in configure.ac and by CMake.

The windows/testc.c file also doesn't need the POSIX `<strings.h>` [1]
header file included as none of the functions defined by it are used:
* ffs
* strcasecmp
* strcasecmp_l
* strncasecmp
* strncasecmp_l

and the `<string.h>` file (part of C89 standard and above) [2] is
sufficient for it.

[1] http://pubs.opengroup.org/onlinepubs/9699919799/basedefs/strings.h.html
[2] https://port70.net/~nsz/c/c89/c89-draft.html#4.1.2